added homepage storybook components
