+++ src/app/sections/Eras.tsx
@@ -1,84 +1,5 @@
 "use client";
-
-import Image from "next/image";
-import Spinner from "../components/spinner/Spinner";
-import { useState } from "react";
-import { motion } from "framer-motion";
-import { twMerge } from "tailwind-merge";
-import { StaticImport } from "next/dist/shared/lib/get-img-props";
-import H1 from "../components/HTML/H1";
-import P from "../components/HTML/P";
-
-function HeroItemCard({
-	title,
-	description,
-	backgroundImage,
-	animateFrom,
-	className,
-}: {
-	title: string;
-	description: string;
-	backgroundImage: StaticImport;
-	animateFrom: "left" | "right" | "bottom";
-	className?: string;
-}) {
-	const [hover, setHover] = useState(false);
-
-	const initialState = {
-		x:
-			animateFrom === "left"
-				? "-100%"
-				: animateFrom === "right"
-					? "100%"
-					: 0,
-		y: "100%",
-		rotate: animateFrom === "left" ? 45 : animateFrom === "right" ? -45 : 0,
-	};
-
-	return (
-		<motion.div
-			animate={{
-				x: 0,
-				y: 0,
-				rotate: 0,
-			}}
-			initial={initialState}
-			transition={{
-				duration: 1,
-				delay: 0.5,
-			}}
-			viewport={{ once: true }}
-			className="relative w-full h-full bg-agblack z-[0] flex justify-end items-start p-[32px] flex-col gap-4"
-			onMouseLeave={() => setHover(false)}
-			onMouseOver={() => setHover(true)}
-			onMouseDown={() => setHover(true)}
-			onMouseMove={() => setHover(true)}
-			onMouseEnter={() => setHover(true)}
-		>
-			<Image
-				src={backgroundImage}
-				alt="Wishing Well"
-				quality={100}
-				height={1136}
-				width={1024}
-				className={twMerge(
-					"absolute top-0 left-0 object-cover h-full w-full -z-10 transition-opacity duration-500",
-					hover ? "opacity-[0.25]" : "opacity-[0.65]",
-					className ?? ""
-				)}
-			/>
-			<H1>{title}</H1>
-			<motion.p
-				animate={{ height: hover ? "auto" : 0 }}
-				initial={{ height: 0 }}
-				transition={{ duration: 0.5 }}
-				className="text-white text-lg overflow-hidden"
-			>
-				<P>{description}</P>
-			</motion.p>
-		</motion.div>
-	);
-}
+import HeroItemCard from "@/stories/HeroItemCard";
 
 export default function Eras() {
 	return (
