+++ src/app/sections/Hero.tsx
@@ -1,84 +1,9 @@
 "use client";
 
-import Image from "next/image";
 import Spinner from "../components/spinner/Spinner";
-import { useEffect, useRef, useState } from "react";
-import { MotionStyle, motion, useScroll, useTransform } from "framer-motion";
-import { twMerge } from "tailwind-merge";
-import { StaticImport } from "next/dist/shared/lib/get-img-props";
-import H1 from "../components/HTML/H1";
-import P from "../components/HTML/P";
-
-function HeroItemCard({
-	title,
-	description,
-	backgroundImage,
-	animateFrom,
-	style,
-}: {
-	title: string;
-	description: string;
-	backgroundImage: StaticImport;
-	style?: React.CSSProperties | MotionStyle;
-	animateFrom: "left" | "right" | "bottom";
-}) {
-	const [hover, setHover] = useState(false);
-
-	const initialState = {
-		x:
-			animateFrom === "left"
-				? "-100%"
-				: animateFrom === "right"
-					? "100%"
-					: 0,
-		y: "100%",
-		rotate: animateFrom === "left" ? 45 : animateFrom === "right" ? -45 : 0,
-	};
-
-	return (
-		<motion.div
-			style={style}
-			animate={{
-				x: 0,
-				y: 0,
-				rotate: 0,
-			}}
-			initial={initialState}
-			transition={{
-				duration: 1,
-				delay: 0.5,
-			}}
-			viewport={{ once: true }}
-			className="relative w-full h-full bg-agblack z-[0] flex justify-end items-start p-[32px] flex-col gap-4"
-			onMouseLeave={() => setHover(false)}
-			onMouseOver={() => setHover(true)}
-			onMouseDown={() => setHover(true)}
-			onMouseMove={() => setHover(true)}
-			onMouseEnter={() => setHover(true)}
-		>
-			<Image
-				src={backgroundImage}
-				alt="Wishing Well"
-				quality={100}
-				height={1136}
-				width={1024}
-				className={twMerge(
-					"absolute top-0 left-0 object-cover h-full w-full -z-10 transition-opacity duration-500",
-					hover ? "opacity-[0.25]" : "opacity-[0.65]"
-				)}
-			/>
-			<H1>{title}</H1>
-			<motion.p
-				animate={{ height: hover ? "auto" : 0 }}
-				initial={{ height: 0 }}
-				transition={{ duration: 0.5 }}
-				className="overflow-hidden"
-			>
-				<P>{description}</P>
-			</motion.p>
-		</motion.div>
-	);
-}
+import { useEffect, useRef } from "react";
+import { motion, useScroll, useTransform } from "framer-motion";
+import HeroItemCard from "@/stories/HeroItemCard";
 
 export default function Hero() {
 	const targetRef = useRef<HTMLDivElement>(null);
@@ -114,21 +39,18 @@ export default function Hero() {
 					description="Contribute to our WishWell to get the WishWell NFT + points."
 					backgroundImage={require("../assets/wishwell.png")}
 					animateFrom="left"
-					style={{}}
 				/>
 				<HeroItemCard
 					title="Mining"
 					description="Start mining with supported tokens to get points + $DARKX tokens + the new Antigravity NFT."
 					backgroundImage={require("../assets/mining.png")}
 					animateFrom="bottom"
-					style={{}}
 				/>
 				<HeroItemCard
 					title="The Collective"
 					description="Learn how to leverage points, rank up & earn exciting rewards. Join The Collective!"
 					backgroundImage={require("../assets/minting.png")}
 					animateFrom="right"
-					style={{}}
 				/>
 			</motion.div>
 		</div>
