+++ src/app/sections/Leaderboard.tsx
@@ -7,210 +7,13 @@ import {
 	motion,
 	AnimatePresence,
 } from "framer-motion";
-import { StaticImport } from "next/dist/shared/lib/get-img-props";
 import Image from "next/image";
 import { useRef, useState } from "react";
-import { twMerge } from "tailwind-merge";
 import H1 from "../components/HTML/H1";
-import H3 from "../components/HTML/H3";
 import H2 from "../components/HTML/H2";
 import P from "../components/HTML/P";
-
-function TH({
-	icon,
-	heading,
-	className,
-}: {
-	icon: string | StaticImport;
-	heading: string;
-	className?: string;
-}) {
-	return (
-		<th
-			className={twMerge(
-				`relative bg-[#0A0025] border-[2px] lg:border-[2px] border-transparent bg-clip-padding flex flex-col lg:flex-row justify-between z-0 px-3 py-[10px] w-full
-            before:content-[''] before:absolute before:inset-0 before:z-[-10] before:bg-gradient-to-bl before:from-[#3C00DC] before:to-[#FF5001] before:rounded-[inherit] before:overflow-hidden before:mb-[-1px] before:ml-[-1px] lg:before:m-[-1px]
-			after:content-[''] after:absolute after:inset-0 after:z-[-2] after:bg-agblack after:rounded-[inherit] after:overflow-hidden`,
-				className
-			)}
-		>
-			<H3>
-				<Image
-					src={icon}
-					alt={`${heading} icon`}
-					width={25}
-					height={25}
-					className="object-cover"
-				/>
-				{heading}
-			</H3>
-		</th>
-	);
-}
-
-function truncatinator(str: string) {
-	return (
-		str.substring(0, 6) + "..." + str.substring(str.length - 4, str.length)
-	);
-}
-
-function TD({
-	children,
-	truncate = false,
-	special = false,
-	className,
-	border = false,
-}: {
-	children: React.ReactNode;
-	truncate?: boolean;
-	special?: boolean;
-	className?: string;
-	border?: boolean;
-}) {
-	const [truncateHover, setTruncateHover] = useState(false);
-	function truncateHoverTrue() {
-		setTruncateHover(true);
-	}
-
-	function truncateHoverFalse() {
-		setTruncateHover(false);
-	}
-
-	if (truncate && children && typeof children === "string") {
-		return (
-			<AnimatePresence>
-				<td
-					onMouseEnter={truncateHoverTrue}
-					onMouseLeave={truncateHoverFalse}
-					className={twMerge(
-						"text-[14px] relative border-r-2 border-[#414343] lg:border-[#8275A5] bg-clip-padding hidden lg:flex flex-col lg:flex-row justify-start items-center z-10 px-[12px] py-[6px] w-full truncate hover:overflow-visible",
-						special && "text-[18px]",
-						className,
-					)}
-				>
-					{truncateHover ? (
-						<>
-							<motion.div
-								animate={{
-									opacity: 1,
-								}}
-								initial={{
-									opacity: 0,
-								}}
-								transition={{ duration: 0.2, delay: 0.5 }}
-								className={twMerge(
-									"absolute top-[50%] left-[50%] p-2 -translate-x-1/2 -translate-y-1/2 rounded-[4px] bg-blue text-center text-agwhite z-10",
-									special &&
-										" text-[18px] bg-agyellow text-black "
-								)}
-							>
-								{children}
-							</motion.div>
-							{truncatinator(children)}
-						</>
-					) : (
-						truncatinator(children)
-					)}
-				</td>
-			</AnimatePresence>
-		);
-	}
-	return (
-		<td
-			className={twMerge(
-				"text-[14px] relative border-r-2 border-[#414343] lg:border-[#8275A5] bg-clip-padding flex flex-col justify-center lg:justify-start lg:items-center gap-[4px] lg:gap-4 lg:flex-row z-0 p-[8px] lg:px-[12px] lg:py-[6px]",
-				special && !border && "border-none text-[18px] py-[10px]",
-				border && special && "border-r-2 text-[18px] py-[10px]",
-				className
-			)}
-		>
-			{children}
-		</td>
-	);
-}
-
-function TR({
-	children,
-	className,
-	special = false,
-	empty = false,
-	th = false,
-}: {
-	children: React.ReactNode;
-	className?: string;
-	special?: boolean;
-	empty?: boolean;
-	th?: boolean;
-}) {
-	return (
-		<tr
-			className={twMerge(
-				"relative grid grid-cols-[2fr_1fr] lg:grid-cols-[2fr_1fr_1fr] w-full lg:border-l-2 border-b-2 border-[#414343] lg:border-[#8275A5] z-0",
-				special && " text-black font-extrabold  border-none",
-				empty &&
-					"bg-gradient-to-b from-[#142266] via-[#0A1133] to-[#142266] border-r-2",
-				th && "border-none",
-				className
-			)}
-		>
-			{special && (
-				<div className="absolute top-0 left-0 w-[calc(100%_+_24px)] h-full translate-x-[-12px] z-[-1] bg-agyellow rounded-lg"></div>
-			)}
-			{children}
-		</tr>
-	);
-}
-
-function Badge({
-	children,
-	special = false,
-}: {
-	children: React.ReactNode;
-	special?: boolean;
-}) {
-	return (
-		<div
-			className={twMerge(
-				"text-[12px] leading-[12px] relative flex items-center gap-[8px] justify-center font-sans font-extrabold text-white cursor-pointer rounded-full py-[4px] px-[8px] border-2 uppercase tracking-widest w-fit",
-				special &&
-					"text-agyellow font-extrabold bg-gradient-to-b from-[#0A1133] to-[#142266] border-none"
-			)}
-		>
-			{children}
-		</div>
-	);
-}
-
-function Rank({
-	rank,
-	wallet,
-	special = false,
-}: {
-	rank: number;
-	wallet: string;
-	special?: boolean;
-}) {
-	return (
-		<TD special={special} border>
-			#{rank} <Badge special={special}>Specialist Technician</Badge>
-			<div className="flex gap-[4px] justify-start items-center lg:hidden text-[18px] lg:text-[14px]">
-				<Image
-					src={require(
-						`@/app/assets/icons/${special ? "wallet-black.svg" : "wallet.svg"}`
-					)}
-					alt="hammer icon"
-					width={25}
-					height={25}
-					className={twMerge(
-						"object-cover",
-						special && "filter-invert"
-					)}
-				/>
-				{truncatinator(wallet)}
-			</div>
-		</TD>
-	);
-}
+import Table from "@/stories/Table";
+import { useAccount } from "wagmi";
 
 function CollectiveLogo() {
 	const [hover, setHover] = useState(false);
@@ -253,10 +56,6 @@ function CollectiveLogo() {
 			</a>
 		</AnimatePresence>
 	);
-}
-
-function pointsConverterToUSCommaseparated(points: number) {
-	return points.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
 }
 
 type tableDataType = {
@@ -325,6 +124,7 @@ export default function Leaderboard({
 }: {
 	accountIsConnected: boolean;
 }) {
+	const account = useAccount();
 	const targetRef = useRef<HTMLDivElement>(null);
 	const { scrollYProgress } = useScroll({
 		target: targetRef,
@@ -349,18 +149,6 @@ export default function Leaderboard({
 				<div className="flex flex-col gap-[16px]">
 					<div className="flex flex-wrap justify-start items-center gap-[16px]">
 						<H1>Leaderboard</H1>
-						{/* <button
-							className={`relative flex items-center gap-x-2 justify-center font-sans uppercase font-extrabold tracking-widest text-agwhite cursor-pointer rounded-lg px-4 py-3 shadow-button shadow-[#414343] border-[#414343] border-2 hover:translate-y-1 transition-[all_150ms] hover:shadow-none active:bg-[rgba(255,255,255,0.25)]`}
-						>
-							<Image
-								src={require("@/app/assets/icons/refresh.svg")}
-								alt="leaderboard icon"
-								width={25}
-								height={25}
-								className="object-cover"
-							/>
-							Refresh
-						</button> */}
 						<Button
 							innerText="Refresh"
 							iconSrc={require("@/app/assets/icons/refresh.svg")}
@@ -373,69 +161,7 @@ export default function Leaderboard({
 
 					<div className="grid grid-cols-1 lg:grid-cols-3 w-full max-w-[100%] pt-6">
 						<div className="col-span-2 w-full rounded-[4px] border-[2px] border-[#414343] lg:border-none">
-							<table className="w-full bg-gradient-to-b from-[#0A1133] to-[#142266] h-fit">
-								<thead className="w-full">
-									<TR th>
-										<TH
-											icon={require("@/app/assets/icons/leaderboard.svg")}
-											heading="Rank"
-										/>
-										<TH
-											icon={require("@/app/assets/icons/wallet.svg")}
-											heading="Wallet"
-											className="hidden lg:flex"
-										/>
-										<TH
-											icon={require("@/app/assets/icons/points.svg")}
-											heading="Points"
-										/>
-									</TR>
-								</thead>
-								<tbody className="text-lg font-medium font-general-sans text-white">
-									{tableData.map((data, idx) =>
-										data !== null ? (
-											<TR
-												key={idx}
-												special={data.special ?? false}
-											>
-												<Rank
-													rank={data.rank}
-													wallet={data.wallet}
-													special={
-														data.special ?? false
-													}
-												/>
-												<TD
-													truncate
-													special={
-														data.special ?? false
-													}
-												>
-													{data.wallet}
-												</TD>
-												<TD
-													special={
-														data.special ?? false
-													}
-													// className=`text-[18px] lg:text-[14px]`
-												>
-													{pointsConverterToUSCommaseparated(
-														data.points
-													)}
-												</TD>
-											</TR>
-										) : (
-											<TR
-												key={idx}
-												className="h-[2.5rem]"
-												empty
-											>
-												<></>
-											</TR>
-										)
-									)}
-								</tbody>
-							</table>
+							<Table tableData={tableData} />
 						</div>
 
 						<div className="relative flex flex-col w-full gap-4 lg:pl-6 place-self-end">
@@ -449,7 +175,7 @@ export default function Leaderboard({
 										wordWrap: "break-word",
 									}}
 								>
-									0x1234567890abcdef1234567890abcdef12345678
+									{account.address}
 								</H2>
 							</div>
 
@@ -467,20 +193,6 @@ export default function Leaderboard({
 									You&apos;re only 1,500 points away from
 									leveling up. Mine now to rank up!
 								</P>
-
-								{/* <button
-									className={`uppercase tracking-widest w-fit relative flex items-center gap-x-2 justify-center font-sans font-bold text-agwhite cursor-pointer
-                                rounded-lg px-4 py-3 shadow-button hover:translate-y-1 transition-[all_150ms] hover:shadow-none active:bg-agblack bg-blue`}
-								>
-									<Image
-										src={require("@/app/assets/icons/hammer.svg")}
-										alt="hammer icon"
-										width={25}
-										height={25}
-										className="object-cover"
-									/>
-									Start mining
-								</button> */}
 								<Button
 									innerText="Start mining"
 									iconSrc={require("@/app/assets/icons/hammer.svg")}
